improvements
Added Russian language support;
English language is separated in another folder;
Changed "Operating on patient" to "Installing ...";
Fxied: Bionic parts now don't appear on the Machining Table. But instead of it, they can be created only on Fabrication Bench as other bionic parts;
Russian language improvement: Make_* added to change "Сделать" (translates to "make" or "do") to "Изготовить" (translates to "manufacture" or "fabricate"). Reason for that is because this technique is used in official russian translation and it is gramatically more correct.

Also I tried to add <descriptionHyperlinks> tag after def's descriptions but looks like all files are located in different folders so... I have no idea :blush: